Move stats panel to appStore
